minor/requirements.txt
# Core Django dependencies
Django==5.0.1
gunicorn==21.2.0
whitenoise==6.6.0

# Database
psycopg2-binary==2.9.9

# Machine Learning
tensorflow-cpu==2.15.0
numpy==1.24.4
Pillow==10.1.0

# API
django-cors-headers==4.3.1

# Static files
django-storages==1.14.2
boto3==1.34.17

# Environment variables
python-dotenv==1.0.0

# ASGI server
uvicorn==0.24.0

# Other utilities
python-multipart==0.0.6
